color ELIF macro
The color ELIF code group macro is complete!

Now the number of colors is not a problem. My macro can take on as many
colors as Excel spreadsheets can have rows. It works by the user pasting
a color table (specifically from this certain website), and at the click
of a button converts the color data into Python ELIF's.

The resulting output files (output.txt) has the Python code generated
from the macro.